added List item

+++ src/main/java/com/henrik/calorieapphenrik/food/dto/FoodDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -13,9 +15,12 @@ public class FoodDto {
 
 
     private String name;
-    private Integer calories;
+    private BigDecimal calories;
     private Integer protein;
     private Integer carbs;
     private Integer fat;
     private Integer fiber;
-}
+
+
+    }
+
